Update from releng/13.1, git revision fc952ac2212.

+++ sys/contrib/openzfs/lib/libshare/os/freebsd/nfs.c
@@ -27,7 +27,7 @@
  */
 
 #include <sys/cdefs.h>
-__FBSDID("$FreeBSD: 5951b9eafa2bc26c4aa9c91f84fb4b47e13a1104 $");
+__FBSDID("$FreeBSD: 97092bdc0fdff5737c83056debae59af2996d178 $");
 
 #include <sys/param.h>
 #include <sys/vfs.h>
@@ -65,17 +65,22 @@ static int nfs_lock_fd = -1;
 static int
 nfs_exports_lock(void)
 {
+	int err;
+
 	nfs_lock_fd = open(ZFS_EXPORTS_LOCK,
-	    O_RDWR | O_CREAT, 0600);
+	    O_RDWR | O_CREAT | O_CLOEXEC, 0600);
 	if (nfs_lock_fd == -1) {
+		err = errno;
 		fprintf(stderr, "failed to lock %s: %s\n",
-		    ZFS_EXPORTS_LOCK, strerror(errno));
-		return (errno);
+		    ZFS_EXPORTS_LOCK, strerror(err));
+		return (err);
 	}
 	if (flock(nfs_lock_fd, LOCK_EX) != 0) {
+		err = errno;
 		fprintf(stderr, "failed to lock %s: %s\n",
-		    ZFS_EXPORTS_LOCK, strerror(errno));
-		return (errno);
+		    ZFS_EXPORTS_LOCK, strerror(err));
+		(void) close(nfs_lock_fd);
+		return (err);
 	}
 	return (0);
 }
@@ -228,8 +233,8 @@ nfs_copy_entries(char *filename, const char *mountpoint)
 	int error = SA_OK;
 	char *line;
 
-	FILE *oldfp = fopen(ZFS_EXPORTS_FILE, "r");
-	FILE *newfp = fopen(filename, "w+");
+	FILE *oldfp = fopen(ZFS_EXPORTS_FILE, "re");
+	FILE *newfp = fopen(filename, "w+e");
 	if (newfp == NULL) {
 		fprintf(stderr, "failed to open %s file: %s", filename,
 		    strerror(errno));
@@ -291,7 +296,7 @@ nfs_enable_share(sa_share_impl_t impl_share)
 		return (error);
 	}
 
-	FILE *fp = fopen(filename, "a+");
+	FILE *fp = fopen(filename, "a+e");
 	if (fp == NULL) {
 		fprintf(stderr, "failed to open %s file: %s", filename,
 		    strerror(errno));
@@ -356,19 +361,15 @@ nfs_disable_share(sa_share_impl_t impl_share)
 	return (error);
 }
 
-/*
- * NOTE: This function returns a static buffer and thus is not thread-safe.
- */
 static boolean_t
 nfs_is_shared(sa_share_impl_t impl_share)
 {
-	static char line[MAXLINESIZE];
-	char *s, last;
+	char *s, last, line[MAXLINESIZE];
 	size_t len;
 	char *mntpoint = impl_share->sa_mountpoint;
 	size_t mntlen = strlen(mntpoint);
 
-	FILE *fp = fopen(ZFS_EXPORTS_FILE, "r");
+	FILE *fp = fopen(ZFS_EXPORTS_FILE, "re");
 	if (fp == NULL)
 		return (B_FALSE);
 
@@ -422,15 +423,21 @@ nfs_commit_shares(void)
 	struct pidfh *pfh;
 	pid_t mountdpid;
 
+start:
 	pfh = pidfile_open(_PATH_MOUNTDPID, 0600, &mountdpid);
 	if (pfh != NULL) {
-		/* Mountd is not running. */
+		/* mountd(8) is not running. */
 		pidfile_remove(pfh);
 		return (SA_OK);
 	}
 	if (errno != EEXIST) {
 		/* Cannot open pidfile for some reason. */
 		return (SA_SYSTEM_ERR);
+	}
+	if (mountdpid == -1) {
+		/* mountd(8) exists, but didn't write the PID yet */
+		usleep(500);
+		goto start;
 	}
 	/* We have mountd(8) PID in mountdpid variable. */
 	kill(mountdpid, SIGHUP);
